+++ src/ionic-tree-provider.ts
@@ -10,7 +10,7 @@ export class IonicTreeProvider implements vscode.TreeDataProvider<Recommendation
 	private _onDidChangeTreeData: vscode.EventEmitter<Recommendation | undefined | void> = new vscode.EventEmitter<Recommendation | undefined | void>();
 	readonly onDidChangeTreeData: vscode.Event<Recommendation | undefined | void> = this._onDidChangeTreeData.event;
 
-	constructor(private workspaceRoot: string | undefined) {
+	constructor(private workspaceRoot: string | undefined, private extensionPath: string) {
 	}
 
 	refresh(): void {
@@ -32,7 +32,7 @@ export class IonicTreeProvider implements vscode.TreeDataProvider<Recommendation
 		} else {
 			const packageJsonPath = path.join(this.workspaceRoot, 'package.json');
 			if (this.pathExists(packageJsonPath)) {
-				return reviewProject(this.workspaceRoot);
+				return reviewProject(this.workspaceRoot, this.extensionPath);
 			} else {
 				return Promise.resolve(starterProject(this.workspaceRoot));
 			}